Basic structure of a task
A task object stores data in a map[string]interface{}. To access the data, the get
and set functions take a string representing a dom path.

A task can be created by parsing JSON data from a string. JSON is also likely to
become the storage format of a task.

The current main reads a string from stdin an tries to parse JSON data out of it.
It then prints the task as go sees is ("%v" format) and exits.